requirements.txt
chromadb
crawl4ai
sentence-transformers
llama-index
llama-index-readers-llama-parse
python-dotenv
pypdf
gradio
duckduckgo_search
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
streamlit
transformers
torch 
llama-index-llms-huggingface
llama-index-llms-ollama
nest-asyncio
fastapi
uvicorn